requirements.txt
Django==3.1.2
requests==2.24.0
djangorestframework==3.12.1
psycopg2~=2.8.6
gunicorn==20.0.4
zeep==4.0.0
lxml==4.5.2
grpcio==1.32.0
marshmallow==3.8.0
pydantic==1.6.1
marshmallow-dataclass[enum,union]
django-cors-headers
djangorestframework-api-key
django_extensions
stripe
haversine
pandas
rauth
django-redis
cryptography
pytz
cachetools==4.1.1